Split motion projection model
