+++ src/app_core/ui_projection/browser_projection/panel.rs
@@ -1,4 +1,5 @@
 use super::*;
+use crate::app_core::state::browser_playback_age_filter_chips;
 use crate::sample_sources::{WavEntry, db::SourceTagUsage};
 
 /// Scalar inputs needed to project the retained browser row window.
